Move fetch page data to app/commands
