+++ web/app/integration-manager/_merlins-connector/app/commands.js
@@ -13,4 +13,3 @@ export const fetchPageData = (url) => (dispatch) => {
             return res
         })
 }
-
